requirements.txt
pytest>=7.4.0
black>=23.3.0
flake8>=6.0.0
python-dotenv>=1.0.0
azure-ai-contentsafety>=1.0.0
requests>=2.31.0
azure-core>=1.29.5
langchain-core>=0.1.27
langgraph>=0.0.15
langchain-openai>=0.0.5
azure-storage-blob>=12.19.0
Pillow>=10.0.0
langsmith>=0.0.77
openmeteo-requests==1.1.0
requests-cache==1.1.1
retry-requests==1.0.0
pandas>=2.0.0